requirements.txt
# 基础依赖
pyyaml>=6.0
requests>=2.28.0

# 日志相关
colorlog>=6.7.0

openai

# 工具相关依赖（可选）
# 如果需要浏览器控制功能，取消下面的注释
# selenium>=4.8.0
# webdriver-manager>=3.8.5